Decompose session runtime helpers
