+++ scripts/audit_maintainability_giant_baseline.toml
@@ -516,9 +516,9 @@
 "src/services/discord/tmux.rs" = 2049
 # completion_guard.rs decomposed in #3479 (1834 -> 872 prod LoC); dropped below
 # the giant threshold, so its ratchet baseline is retired (no longer tracked).
-# #3034 batch-8: lowered 1781 -> 1753 (deleted the dead `assistant_turns` turn-cap
-# tracker — field + 3 methods + construction sites).
-"src/services/discord/session_runtime.rs" = 1753
+# session_runtime.rs decomposed in #3842 (worktree, restore_cwd, and
+# channel_routing child modules); dropped below the giant threshold, so its
+# ratchet baseline is retired.
 # #3034 batch-8: +1 for a test-only `#[allow(dead_code)]` on
 # `session_bound_terminal_delivery_route_decision` (#3041 route decision, tested).
 # #3610 PR-1: 1731 -> 1738 — pass terminal anchor (current_msg_id) explicitly to
