+++ src/services/discord/session_runtime.rs
@@ -1,6 +1,34 @@
-use super::runtime_store::worktrees_root;
 use super::*;
-use crate::services::git::GitCommand;
+
+#[path = "session_runtime/channel_routing.rs"]
+mod channel_routing;
+#[path = "session_runtime/restore_cwd.rs"]
+mod restore_cwd;
+#[path = "session_runtime/worktree.rs"]
+mod worktree;
+
+use self::channel_routing::choose_restore_channel_name;
+pub(super) use self::channel_routing::{
+    RuntimeChannelBindingStatus, provider_handles_channel, resolve_channel_category,
+    resolve_is_dm_channel, resolve_runtime_channel_binding_status, resolve_thread_parent,
+    synthetic_thread_channel_name, validate_live_channel_routing,
+    validate_live_channel_routing_with_dm_hint,
+};
+#[cfg(test)]
+use self::restore_cwd::restore_thread_worktree_path_from_db;
+use self::restore_cwd::{
+    correct_session_cwd_to_tmux, reconcile_recovery_cwd, restore_session_cwd_from_db,
+};
+pub(super) use self::restore_cwd::{
+    db_cwd_is_reusable_worktree, select_restored_session_path, session_path_is_usable,
+};
+pub(super) use self::worktree::{
+    WorktreeInfo, cleanup_git_worktree, create_git_worktree, detect_worktree_conflict,
+    resolve_reusable_worktree,
+};
+use self::worktree::{
+    is_managed_worktree_path, reconstruct_managed_worktree_metadata, sync_inflight_worktree_context,
+};
 
 /// Per-channel session state
 #[derive(Clone)]
@@ -27,98 +55,6 @@ pub(super) struct DiscordSession {
     /// Restart generation at which this session was created/restored.
     #[allow(dead_code)]
     pub(super) born_generation: u64,
-}
-
-pub(super) fn session_path_is_usable(current_path: &str) -> bool {
-    std::path::Path::new(current_path).is_dir()
-}
-
-pub(super) fn select_restored_session_path(
-    configured_path: Option<String>,
-    db_cwd: Option<String>,
-    yaml_path: Option<String>,
-    db_cwd_is_reusable_worktree: bool,
-) -> Option<String> {
-    // #3219: when the channel-scoped DB cwd is the channel's OWN existing managed
-    // worktree (caller-validated by `db_cwd_is_reusable_worktree`: under
-    // `worktrees_root`, a linked worktree, and sharing the configured parent
-    // repo's git common dir), prefer it over the configured *base* workspace.
-    // Otherwise crash/kill recovery installs the base as cwd, provider-channel
-    // worktree isolation re-derives a FRESH worktree + provider session-id, and
-    // `--resume` breaks — abandoning the session's transcript. The live-TUI
-    // -binding recovery path masks this only while the tmux pane survives; once
-    // the pane dies there is no fallback (root cause of the 2026-06-07 resume
-    // failure: recovery read the correct worktree from the DB, logged "Ignoring
-    // restored DB cwd", then built a fresh worktree + session-id).
-    //
-    // The predicate uses the SAME guard set as `resolve_reusable_worktree`, so a
-    // stale/foreign/relocated worktree — including a workspace repointed to a
-    // different repo under the same `worktrees_root` — is NOT elevated and falls
-    // through to the configured path below.
-    if db_cwd_is_reusable_worktree
-        && let Some(worktree) = db_cwd.as_ref().filter(|path| session_path_is_usable(path))
-    {
-        return Some(worktree.clone());
-    }
-
-    configured_path
-        .filter(|path| session_path_is_usable(path))
-        .or_else(|| db_cwd.filter(|path| session_path_is_usable(path)))
-        .or_else(|| yaml_path.filter(|path| session_path_is_usable(path)))
-}
-
-/// #3219: true when the recovered DB cwd is the channel's own reusable managed
-/// worktree and must therefore outrank the configured base workspace in
-/// [`select_restored_session_path`]. Mirrors the exact guard set used by
-/// [`resolve_reusable_worktree`]: the cwd must be an AgentDesk-managed
-/// (`is_managed_worktree_path`) linked worktree that shares the configured
-/// parent repo's git common dir (`restored_worktree_belongs_to_parent`, which
-/// also rejects non-existent and remote-only paths). Returns `false` when there
-/// is no configured parent (then `select_restored_session_path`'s existing
-/// configured→db_cwd→yaml fallback already does the right thing).
-pub(super) fn db_cwd_is_reusable_worktree(
-    configured_path: Option<&str>,
-    db_cwd: Option<&str>,
-) -> bool {
-    match (configured_path, db_cwd) {
-        (Some(parent), Some(cwd)) => {
-            is_managed_worktree_path(cwd) && restored_worktree_belongs_to_parent(parent, cwd)
-        }
-        _ => false,
-    }
-}
-
-/// #3216 GAP 2: decide whether a live tmux pane's cwd should override the
-/// DB-resolved cwd during recovery. The live tmux pane is the authoritative
-/// source of truth for where a session is actually running — if the DB cwd has
-/// diverged (e.g. a phantom worktree rotation stamped a transcript-less path),
-/// trusting the DB blindly relaunches `--resume` against the wrong cwd and the
-/// conversation is lost.
-///
-/// Returns `Some(tmux_cwd)` only when ALL hold:
-///   * a live tmux pane cwd is present (`tmux_cwd`);
-///   * it is a real AgentDesk-managed, on-disk-usable worktree
-///     (guarded by `tmux_cwd_is_managed` / `tmux_cwd_is_usable` predicates so we
-///     never adopt a transient or garbage path);
-///   * it actually DIFFERS from the DB cwd (nothing to reconcile otherwise).
-///
-/// Kept as a pure function (predicate results injected) so the reconcile policy
-/// is unit-testable without a live tmux / filesystem.
-fn reconcile_recovery_cwd(
-    db_cwd: Option<&str>,
-    tmux_cwd: Option<&str>,
-    tmux_cwd_is_managed: bool,
-    tmux_cwd_is_usable: bool,
-) -> Option<String> {
-    let tmux_cwd = tmux_cwd?.trim();
-    if tmux_cwd.is_empty() || !tmux_cwd_is_managed || !tmux_cwd_is_usable {
-        return None;
-    }
-    if db_cwd.map(str::trim) == Some(tmux_cwd) {
-        // DB already agrees with the live pane — nothing to reconcile.
-        return None;
-    }
-    Some(tmux_cwd.to_string())
 }
 
 impl DiscordSession {
@@ -168,427 +104,6 @@ pub(super) fn restored_memento_context_loaded(
     next_session_id: Option<&str>,
 ) -> bool {
     previous_loaded && previous_session_id == next_session_id && next_session_id.is_some()
-}
-
-/// Worktree info for sessions that were auto-redirected to avoid conflicts
-#[derive(Clone, Debug)]
-pub(super) struct WorktreeInfo {
-    /// The original repo path that was conflicted
-    pub original_path: String,
-    /// The worktree directory path
-    pub(super) worktree_path: String,
-    /// The branch name created for this worktree
-    pub(super) branch_name: String,
-}
-
-pub(super) fn synthetic_thread_channel_name(parent_name: &str, channel_id: ChannelId) -> String {
-    format!("{parent_name}-t{}", channel_id.get())
-}
-
-pub(super) fn is_synthetic_thread_channel_name(channel_name: &str, channel_id: ChannelId) -> bool {
-    channel_name.ends_with(&format!("-t{}", channel_id.get()))
-}
-
-pub(super) fn choose_restore_channel_name(
-    existing_channel_name: Option<&str>,
-    live_channel_name: Option<&str>,
-    thread_parent: Option<(ChannelId, Option<String>)>,
-    channel_id: ChannelId,
-) -> Option<String> {
-    if let Some(existing_name) = existing_channel_name
-        && is_synthetic_thread_channel_name(existing_name, channel_id)
-    {
-        return Some(existing_name.to_string());
-    }
-
-    if let Some((parent_id, parent_name)) = thread_parent {
-        let parent_name = parent_name.unwrap_or_else(|| parent_id.get().to_string());
-        return Some(synthetic_thread_channel_name(&parent_name, channel_id));
-    }
-
-    live_channel_name
-        .or(existing_channel_name)
-        .map(ToOwned::to_owned)
-}
-
-pub(super) fn resolve_is_dm_channel(
-    dm_hint: Option<bool>,
-    live_channel_lookup_says_dm: bool,
-) -> bool {
-    // Prefer the gateway-provided DM hint when available so a transient
-    // Discord channel lookup failure cannot disable DM default-agent fallback.
-    dm_hint.unwrap_or(live_channel_lookup_says_dm)
-}
-
-/// Check if a path is a git repo and if another channel already uses it.
-/// Returns the conflicting channel's name if found.
-pub(super) fn detect_worktree_conflict(
-    sessions: &HashMap<ChannelId, DiscordSession>,
-    path: &str,
-    my_channel: ChannelId,
-) -> Option<String> {
-    let norm = path.trim_end_matches('/');
-    for (cid, session) in sessions {
-        if *cid == my_channel {
-            continue;
-        }
-        let other_path = if let Some(ref wt) = session.worktree {
-            &wt.original_path
-        } else {
-            match &session.current_path {
-                Some(p) => p.as_str(),
-                None => continue,
-            }
-        };
-        if other_path.trim_end_matches('/') == norm {
-            return session
-                .channel_name
-                .clone()
-                .or_else(|| Some(cid.get().to_string()));
-        }
-    }
-    None
-}
-
-/// Create a git worktree for the given repo path.
-/// Returns (worktree_path, branch_name) on success.
-pub(super) fn create_git_worktree(
-    repo_path: &str,
-    channel_name: &str,
-    provider: &str,
-) -> Result<(String, String), String> {
-    if GitCommand::new()
-        .repo(repo_path)
-        .args(["rev-parse", "--is-inside-work-tree"])
-        .run_output()
-        .is_err()
-    {
-        return Err(format!("{} is not a git repository", repo_path));
-    }
-
-    let ts = chrono::Local::now().format("%Y%m%d-%H%M%S");
-    let safe_name = channel_name
-        .chars()
-        .map(|c| {
-            if c.is_alphanumeric() || c == '-' {
-                c
-            } else {
-                '_'
-            }
-        })
-        .collect::<String>();
-    let branch = format!("wt/{}-{}-{}", provider, safe_name, ts);
-
-    let wt_base = worktrees_root().ok_or("Cannot determine worktree root")?;
-    std::fs::create_dir_all(&wt_base)
-        .map_err(|e| format!("Failed to create worktree base dir: {}", e))?;
-    let wt_dir = wt_base.join(format!("{}-{}-{}", provider, safe_name, ts));
-    let wt_path = wt_dir.display().to_string();
-    let base_ref = git_upstream_base_ref(repo_path);
-
-    GitCommand::new()
-        .repo(repo_path)
-        .args(["worktree", "add", "-b"])
-        .arg(&branch)
-        .arg(&wt_path)
-        .arg(&base_ref)
-        .run_output()
-        .map_err(|e| format!("git worktree add failed: {}", e))?;
-
-    let ts_log = chrono::Local::now().format("%H:%M:%S");
-    tracing::info!(
-        "  [{ts_log}] 🌿 Created worktree: {} (branch: {})",
-        wt_path,
-        branch
-    );
-    Ok((wt_path, branch))
-}
-
-fn git_upstream_base_ref(repo_path: &str) -> String {
-    if GitCommand::new()
-        .repo(repo_path)
-        .args(["rev-parse", "--verify", "origin/main"])
-        .run_output()
-        .is_ok()
-    {
-        return "origin/main".to_string();
-    }
-
-    // origin/main not available locally — attempt a shallow fetch before falling back
-    if GitCommand::new()
-        .repo(repo_path)
-        .args(["fetch", "origin", "main", "--depth=1"])
-        .run_output()
-        .is_ok()
-    {
-        // Re-verify after fetch
-        if GitCommand::new()
-            .repo(repo_path)
-            .args(["rev-parse", "--verify", "origin/main"])
-            .run_output()
-            .is_ok()
-        {
-            tracing::info!(
-                "git fetch origin main --depth=1 succeeded for repo {repo_path}; using origin/main as base ref"
-            );
-            return "origin/main".to_string();
-        }
-    }
-
-    tracing::warn!(
-        "origin/main unavailable for repo {repo_path} even after fetch attempt; falling back to local 'main'"
-    );
-    "main".to_string()
-}
-
-fn worktree_has_local_changes(wt_info: &WorktreeInfo) -> Result<bool, String> {
-    let status = GitCommand::new()
-        .repo(&wt_info.worktree_path)
-        .args(["status", "--porcelain"])
-        .run_output()
-        .map_err(|e| format!("git status failed: {e}"))?;
-    Ok(!status.stdout.is_empty())
-}
-
-fn git_command_output(repo_path: &str, args: &[&str]) -> Result<std::process::Output, String> {
-    GitCommand::new()
-        .repo(repo_path)
-        .args(args.iter().copied())
-        .run_output()
-        .map_err(|e| format!("git {:?} failed: {e}", args))
-}
-
-fn git_command_stdout(repo_path: &str, args: &[&str]) -> Result<String, String> {
-    let output = git_command_output(repo_path, args)?;
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-        return Err(format!("git {:?} failed: {stderr}", args));
-    }
-    Ok(String::from_utf8_lossy(&output.stdout).trim().to_string())
-}
-
-fn patch_id_from_diff(diff: &[u8]) -> Result<Option<String>, String> {
-    if diff.is_empty() {
-        return Ok(None);
-    }
-
-    let output = GitCommand::new()
-        .args(["patch-id", "--stable"])
-        .run_output_with_stdin(diff)
-        .map_err(|e| format!("git patch-id failed: {e}"))?;
-
-    Ok(String::from_utf8_lossy(&output.stdout)
-        .split_whitespace()
-        .next()
-        .map(str::to_string))
-}
-
-fn branch_diff_patch_id(
-    repo_path: &str,
-    from_ref: &str,
-    to_ref: &str,
-) -> Result<Option<String>, String> {
-    let range = format!("{from_ref}..{to_ref}");
-    let diff = git_command_output(repo_path, &["diff", "--binary", "--no-ext-diff", &range])?;
-    if !diff.status.success() {
-        let stderr = String::from_utf8_lossy(&diff.stderr).trim().to_string();
-        return Err(format!("git diff {range} failed: {stderr}"));
-    }
-    patch_id_from_diff(&diff.stdout)
-}
-
-fn commit_patch_id(repo_path: &str, commit_sha: &str) -> Result<Option<String>, String> {
-    let show = git_command_output(
-        repo_path,
-        &[
-            "show",
-            "--format=",
-            "--patch",
-            "--binary",
-            "--no-ext-diff",
-            commit_sha,
-        ],
-    )?;
-    if !show.status.success() {
-        let stderr = String::from_utf8_lossy(&show.stderr).trim().to_string();
-        return Err(format!("git show {commit_sha} failed: {stderr}"));
-    }
-    patch_id_from_diff(&show.stdout)
-}
-
-fn worktree_is_squash_merged(
-    repo_path: &str,
-    base_ref: &str,
-    branch_name: &str,
-) -> Result<bool, String> {
-    let merge_base = git_command_stdout(repo_path, &["merge-base", base_ref, branch_name])?;
-    if merge_base.is_empty() {
-        return Ok(false);
-    }
-
-    let Some(branch_patch_id) = branch_diff_patch_id(repo_path, &merge_base, branch_name)? else {
-        return Ok(false);
-    };
-
-    let commit_range = format!("{merge_base}..{base_ref}");
-    let base_commits = git_command_stdout(repo_path, &["rev-list", "--no-merges", &commit_range])?;
-    for commit_sha in base_commits
-        .lines()
-        .map(str::trim)
-        .filter(|sha| !sha.is_empty())
-    {
-        if commit_patch_id(repo_path, commit_sha)?.as_deref() == Some(branch_patch_id.as_str()) {
-            tracing::info!(
-                "Detected squash-merged worktree branch {branch_name} via patch-id match on {commit_sha}"
-            );
-            return Ok(true);
-        }
-    }
-
-    Ok(false)
-}
-
-fn disconnect_sessions_for_worktree_path(
-    _db: Option<&crate::db::Db>,
-    pg_pool: Option<&sqlx::PgPool>,
-    worktree_path: &str,
-) {
-    if let Some(pool) = pg_pool {
-        let worktree_path_owned = worktree_path.to_string();
-        match crate::utils::async_bridge::block_on_pg_result(
-            pool,
-            move |bridge_pool| async move {
-                let updated = sqlx::query(
-                    "UPDATE sessions
-                     SET cwd = NULL,
-                         status = 'disconnected',
-                         active_dispatch_id = NULL,
-                         claude_session_id = NULL
-                     WHERE cwd = $1",
-                )
-                .bind(&worktree_path_owned)
-                .execute(&bridge_pool)
-                .await
-                .map_err(|err| {
-                    format!(
-                        "disconnect pg sessions for removed worktree {worktree_path_owned}: {err}"
-                    )
-                })?
-                .rows_affected();
-                Ok(updated)
-            },
-            |error| error,
-        ) {
-            Ok(updated) if updated > 0 => tracing::info!(
-                "Disconnected {updated} PG session(s) referencing removed worktree {}",
-                worktree_path
-            ),
-            Ok(_) => {}
-            Err(err) => tracing::warn!(
-                "Failed to disconnect PG sessions for removed worktree {}: {}",
-                worktree_path,
-                err
-            ),
-        }
-    }
-}
-
-fn worktree_has_unmerged_commits(wt_info: &WorktreeInfo) -> Result<bool, String> {
-    let base_ref = git_upstream_base_ref(&wt_info.original_path);
-    let range = format!("{base_ref}..{}", wt_info.branch_name);
-    let diff = git_command_output(
-        &wt_info.original_path,
-        &["log", "--oneline", range.as_str()],
-    )?;
-    if !diff.status.success() {
-        let stderr = String::from_utf8_lossy(&diff.stderr).trim().to_string();
-        return Err(format!("git log failed: {stderr}"));
-    }
-    if diff.stdout.is_empty() {
-        return Ok(false);
-    }
-
-    if worktree_is_squash_merged(&wt_info.original_path, &base_ref, &wt_info.branch_name)? {
-        return Ok(false);
-    }
-
-    Ok(true)
-}
-
-/// Clean up a git worktree after session ends.
-pub(super) fn cleanup_git_worktree(
-    db: Option<&crate::db::Db>,
-    pg_pool: Option<&sqlx::PgPool>,
-    wt_info: &WorktreeInfo,
-) {
-    let ts = chrono::Local::now().format("%H:%M:%S");
-
-    let has_changes = match worktree_has_local_changes(wt_info) {
-        Ok(value) => value,
-        Err(err) => {
-            tracing::warn!(
-                "  [{ts}] ⚠ Could not inspect worktree {} for cleanup: {} — preserving",
-                wt_info.worktree_path,
-                err
-            );
-            return;
-        }
-    };
-
-    let has_commits = match worktree_has_unmerged_commits(wt_info) {
-        Ok(value) => value,
-        Err(err) => {
-            tracing::warn!(
-                "  [{ts}] ⚠ Could not inspect branch {} for cleanup: {} — preserving",
-                wt_info.branch_name,
-                err
-            );
-            return;
-        }
-    };
-
-    if has_changes || has_commits {
-        tracing::info!(
-            "  [{ts}] 🌿 Worktree {} has changes/unmerged commits — preserving until merge cleanup",
-            wt_info.worktree_path
-        );
-        tracing::info!(
-            "  [{ts}] 🌿 Branch: {} | Original: {}",
-            wt_info.branch_name,
-            wt_info.original_path
-        );
-    } else {
-        if let Err(err) = GitCommand::new()
-            .repo(&wt_info.original_path)
-            .args(["worktree", "remove"])
-            .arg(&wt_info.worktree_path)
-            .run_output()
-        {
-            tracing::warn!(
-                "  [{ts}] ⚠ Failed to remove worktree {}: {} — preserving DB session path",
-                wt_info.worktree_path,
-                err
-            );
-            return;
-        }
-
-        let branch_delete = GitCommand::new()
-            .repo(&wt_info.original_path)
-            .args(["branch", "-D"])
-            .arg(&wt_info.branch_name)
-            .run_output();
-        let _ = std::fs::remove_dir_all(&wt_info.worktree_path);
-        disconnect_sessions_for_worktree_path(db, pg_pool, &wt_info.worktree_path);
-        if let Err(err) = branch_delete {
-            tracing::warn!(
-                "  [{ts}] ⚠ Removed worktree {} but could not delete branch {}: {}",
-                wt_info.worktree_path,
-                wt_info.branch_name,
-                err
-            );
-        }
-        tracing::info!("  [{ts}] 🧹 Cleaned up worktree: {}", wt_info.worktree_path);
-    }
 }
 
 /// Auto-restore session from bot_settings.json if not in memory
@@ -851,469 +366,6 @@ pub(super) async fn auto_restore_session_force(
     }
 }
 
-/// Resolve a channel's persisted `sessions.cwd` for the given `session_key`,
-/// scoped to the unique Discord `channel_id`, with a SAFE legacy fallback for
-/// rows that predate the `channel_id` column.
-///
-/// #3216 GAP 1: migration `0071_sessions_channel_id.sql` only added the column;
-/// existing rows kept `channel_id = NULL`. The strict `channel_id = $2` guard
-/// (#3207) therefore never matches a legacy row, so the FIRST restore after
-/// deploy still rotates a brand-new worktree and divorces the live session from
-/// its transcript. We cannot backfill the numeric id in pure SQL (the id is not
-/// derivable from `session_key`, which holds the channel NAME), so instead we
-/// fall back to a name-only lookup — but ONLY when it is unambiguous:
-///
-///   * channel-scoped match (`session_key = $1 AND channel_id = $2`) wins first;
-///   * otherwise, fall back ONLY when there is EXACTLY ONE row for the
-///     `session_key` AND that row's `channel_id IS NULL` (a true legacy row).
-///
-/// `sessions.session_key` is globally UNIQUE (migration `001_initial.sql`), so
-/// there is at most one row per key; the `rows.len() == 1` check is therefore a
-/// defensive belt-and-braces guard. If the single row carries a DIFFERENT
-/// non-null `channel_id`, the fallback is refused — that would reintroduce the
-/// #3207 cross-channel hazard. New heartbeats stamp `channel_id`, so this
-/// self-heals over time. Returns a [`RestoredCwd`] for this `session_key`:
-/// `channel_scoped = true` for an exact channel-id match (whose `path` may be
-/// empty when the owned row's `cwd` is NULL/missing — ownership comes from row
-/// existence, #3219), or `channel_scoped = false` with a non-empty path for the
-/// legacy NULL fallback.
-async fn resolve_cwd_for_session_key(
-    pool: &sqlx::PgPool,
-    session_key: &str,
-    channel_id: &str,
-) -> Result<Option<RestoredCwd>, String> {
-    // 1. Channel-scoped match (the #3207 cross-channel guard). `fetch_optional`
-    //    returns the OUTER Option: `Some(_)` iff an exact channel-owned row
-    //    exists, independent of whether its `cwd` is currently populated. #3219:
-    //    ownership is reported (`channel_scoped: true`) from row EXISTENCE — a
-    //    NULL/empty cwd must NOT erase ownership, or an owned channel whose row
-    //    has a stale/missing cwd could not elevate the valid worktree the tmux
-    //    reconcile later supplies. The caller filters the (possibly empty) path
-    //    for usability separately.
-    let scoped = sqlx::query_scalar::<_, Option<String>>(
-        "SELECT cwd FROM sessions \
-         WHERE session_key = $1 AND channel_id = $2 LIMIT 1",
-    )
-    .bind(session_key)
-    .bind(channel_id)
-    .fetch_optional(pool)
-    .await
-    .map_err(|error| format!("load session cwd {session_key}: {error}"))?;
-    if let Some(cwd) = scoped {
-        return Ok(Some(RestoredCwd {
-            path: cwd.unwrap_or_default(),
-            channel_scoped: true,
-        }));
-    }
-
-    // 2. #3216 GAP 1 safe legacy fallback: inspect ALL rows for this
-    //    session_key. Honor only the unambiguous single-NULL-channel_id case.
-    let rows = sqlx::query_as::<_, (Option<String>, Option<String>)>(
-        "SELECT cwd, channel_id FROM sessions WHERE session_key = $1",
-    )
-    .bind(session_key)
-    .fetch_all(pool)
-    .await
-    .map_err(|error| format!("load legacy session cwd {session_key}: {error}"))?;
-
-    if rows.len() == 1 {
-        let (cwd, row_channel_id) = &rows[0];
-        if row_channel_id.is_none()
-            && let Some(path) = cwd.clone().filter(|p| !p.is_empty())
-        {
-            tracing::info!(
-                "  ↻ #3216 legacy NULL-channel_id fallback: reusing cwd {} for \
-                 session_key={} (channel_id={}); row will self-heal on next heartbeat",
-                path,
-                session_key,
-                channel_id
-            );
-            // #3219: a NULL-channel_id row is NOT proven to belong to THIS
-            // channel (a name-collision channel resolves the same globally-unique
-            // session_key). Mark it non-channel-scoped so it never gets elevated
-            // over the safe configured base in `select_restored_session_path`.
-            return Ok(Some(RestoredCwd {
-                path,
-                channel_scoped: false,
-            }));
-        }
-    }
-    Ok(None)
-}
-
-/// Resolve a channel's persisted `sessions.cwd` for restart restore, scoped to
-/// the unique Discord `channel_id` (with the #3216 safe legacy fallback).
-///
-/// Backs the `db_cwd` lookup in [`auto_restore_session_force`], which installs
-/// the resolved path into `session.current_path`. See
-/// [`resolve_cwd_for_session_key`] for the channel-scoping / legacy-fallback
-/// semantics.
-///
-/// The on-disk usability filter (`session_path_is_usable`) is applied by the
-/// caller so this helper stays a pure DB resolve.
-/// A persisted `sessions.cwd` resolved during restart recovery, tagged with
-/// whether it came from an exact channel-id match (`channel_scoped = true`) or
-/// the #3216 legacy NULL-channel_id fallback (`false`). Only a channel-scoped
-/// cwd is eligible to outrank the configured base in
-/// [`select_restored_session_path`] (#3219) — a NULL-fallback cwd is not proven
-/// to belong to this channel.
-#[derive(Debug, Clone)]
-struct RestoredCwd {
-    path: String,
-    channel_scoped: bool,
-}
-
-fn restore_session_cwd_from_db(
-    pg_pool: Option<&sqlx::PgPool>,
-    token_hash: &str,
-    provider: &ProviderKind,
-    channel_name: &str,
-    channel_id: u64,
-) -> Option<RestoredCwd> {
-    let tmux_name = provider.build_tmux_session_name(channel_name);
-    let session_keys = build_session_key_candidates(token_hash, provider, &tmux_name);
-    let channel_id = channel_id.to_string();
-    let pg_pool = pg_pool?;
-    crate::utils::async_bridge::block_on_pg_result(
-        pg_pool,
-        move |pool| async move {
-            for session_key in session_keys {
-                if let Some(restored) =
-                    resolve_cwd_for_session_key(&pool, &session_key, &channel_id).await?
-                {
-                    return Ok(Some(restored));
-                }
-            }
-            Ok(None)
-        },
-        |message| message,
-    )
-    .ok()
-    .flatten()
-}
-
-/// #3216 GAP 2: correct the persisted `sessions.cwd` to the authoritative live
-/// tmux pane cwd during recovery reconciliation. Scoped by `session_key` AND the
-/// unique `channel_id` so a name collision can never write into another channel's
-/// row (the #3207 cross-channel guard). A legacy NULL-channel_id row is matched
-/// too — only the row whose id equals THIS channel OR is NULL is updated — so the
-/// row both adopts the correct cwd and gets self-healed onto this channel.
-///
-fn correct_session_cwd_to_tmux(
-    pg_pool: Option<&sqlx::PgPool>,
-    token_hash: &str,
-    provider: &ProviderKind,
-    channel_name: &str,
-    channel_id: u64,
-    tmux_cwd: &str,
-) {
-    let Some(pool) = pg_pool else {
-        return;
-    };
-    let tmux_name = provider.build_tmux_session_name(channel_name);
-    let session_keys = build_session_key_candidates(token_hash, provider, &tmux_name);
-    let channel_id_str = channel_id.to_string();
-    let tmux_cwd = tmux_cwd.to_string();
-    let result = crate::utils::async_bridge::block_on_pg_result(
-        pool,
-        move |bridge_pool| async move {
-            let mut total = 0u64;
-            for session_key in session_keys {
-                let updated = sqlx::query(
-                    "UPDATE sessions \
-                     SET cwd = $1, channel_id = $2 \
-                     WHERE session_key = $3 \
-                       AND (channel_id = $2 OR channel_id IS NULL)",
-                )
-                .bind(&tmux_cwd)
-                .bind(&channel_id_str)
-                .bind(&session_key)
-                .execute(&bridge_pool)
-                .await
-                .map_err(|err| format!("reconcile cwd for {session_key}: {err}"))?
-                .rows_affected();
-                total += updated;
-            }
-            Ok(total)
-        },
-        |error| error,
-    );
-    match result {
-        Ok(updated) if updated > 0 => tracing::info!(
-            "  ↻ #3216 reconciled DB cwd to live tmux pane for channel {} ({} row(s))",
-            channel_id,
-            updated
-        ),
-        Ok(_) => {}
-        Err(err) => tracing::warn!(
-            "  ⚠ #3216 failed to reconcile DB cwd to live tmux for channel {}: {}",
-            channel_id,
-            err
-        ),
-    }
-}
-
-/// Look up the persisted worktree path for a thread session from the `sessions`
-/// DB table, mirroring the restore lookup in [`auto_restore_session_force`].
-///
-/// After a dcserver restart the in-memory `sessions` map is empty, so without
-/// this lookup a new thread message would create a brand-new worktree and drop
-/// the provider session fingerprint / recovery context tied to the previous
-/// worktree path (#3011). The returned path is only honored when it still names
-/// a usable git worktree on disk; otherwise we fall back to creating a fresh one.
-///
-/// #3207 (part 2) P0: the `session_key` is derived from the sanitized/truncated
-/// channel NAME, so two distinct channels whose names collide produce the SAME
-/// `session_key` and would resolve EACH OTHER's persisted cwd. The lookup is
-/// therefore scoped by the unique `channel_id` — only a row stamped with THIS
-/// channel's id is honored, so a name collision can never cross channels.
-///
-/// #3216 GAP 1: legacy rows written before the `channel_id` column existed carry
-/// `channel_id = NULL` and so never match the strict scoped predicate, forcing a
-/// worktree rotation on the first restore after deploy. [`resolve_cwd_for_session_key`]
-/// adds a SAFE fallback that reuses such a row ONLY when it is unambiguous
-/// (exactly one row for the `session_key` and its `channel_id IS NULL`), which
-/// preserves the cross-channel guard while letting legacy sessions keep their
-/// transcript-bearing worktree.
-fn restore_thread_worktree_path_from_db(
-    pg_pool: Option<&sqlx::PgPool>,
-    token_hash: &str,
-    provider: &ProviderKind,
-    channel_name: &str,
-    channel_id: u64,
-) -> Option<String> {
-    let tmux_name = provider.build_tmux_session_name(channel_name);
-    let session_keys = build_session_key_candidates(token_hash, provider, &tmux_name);
-    let channel_id = channel_id.to_string();
-    let pg_pool = pg_pool?;
-    crate::utils::async_bridge::block_on_pg_result(
-        pg_pool,
-        move |pool| async move {
-            for session_key in session_keys {
-                if let Some(restored) =
-                    resolve_cwd_for_session_key(&pool, &session_key, &channel_id).await?
-                {
-                    // The worktree-reuse path applies its own managed/belongs-to
-                    // -parent guards downstream; it only needs the path here. Skip
-                    // an owned row whose cwd is empty/NULL (#3219) — it carries no
-                    // reusable worktree, so continue scanning the remaining keys.
-                    if !restored.path.is_empty() {
-                        return Ok(Some(restored.path));
-                    }
-                }
-            }
-            Ok(None)
-        },
-        |message| message,
-    )
-    .ok()
-    .flatten()
-}
-
-/// True when `worktree_path` exists, is a git worktree, and shares the same
-/// repository (git common dir) as `parent_path`. The shared-repo match is
-/// essential: a thread can be reused by a later dispatch that targets a
-/// *different* repo, in which case the stored cwd must NOT be restored — we must
-/// fall through to create a worktree off the requested `parent_path` so the
-/// dispatch runs against its real target (#3011 codex review: avoid treating a
-/// restored cwd as the dispatch target).
-///
-/// Both paths are compared by their `--git-common-dir` so the check holds even
-/// when `parent_path` is itself a linked worktree (e.g. a dispatch worktree),
-/// where comparing against the main checkout would otherwise reject a valid
-/// restored thread worktree.
-fn restored_worktree_belongs_to_parent(parent_path: &str, worktree_path: &str) -> bool {
-    if !std::path::Path::new(worktree_path).is_dir() {
-        return false;
-    }
-    // Only accept a *distinct linked* worktree. If a previous
-    // `create_git_worktree` failure persisted the fallback `parent_path` (the
-    // shared parent checkout) as this thread's cwd, restoring it would record
-    // the main checkout as `session.worktree`, defeating isolation and exposing
-    // it to worktree idle-cleanup. A linked worktree's per-worktree git dir
-    // differs from its shared common dir; the main checkout's do not.
-    if !is_linked_worktree(worktree_path) {
-        return false;
-    }
-    let worktree_repo = match git_common_dir(worktree_path) {
-        Some(dir) => dir,
-        None => return false,
-    };
-    let parent_repo = match git_common_dir(parent_path) {
-        Some(dir) => dir,
-        None => return false,
-    };
-    paths_equal(&worktree_repo, &parent_repo)
-}
-
-/// True when `path` lives under the AgentDesk-managed worktrees root, i.e. it is
-/// a worktree this process created via [`create_git_worktree`] and therefore owns
-/// for cleanup. A user's *configured workspace* that happens to be a linked git
-/// worktree lives elsewhere and must NOT be treated as a disposable
-/// AgentDesk-owned worktree — otherwise idle session cleanup could remove the
-/// user's checkout and delete its branch (#3011 codex review P1).
-fn is_managed_worktree_path(path: &str) -> bool {
-    let Some(root) = worktrees_root() else {
-        return false;
-    };
-    let root = root.canonicalize().unwrap_or(root);
-    let candidate = std::path::PathBuf::from(path);
-    let candidate = candidate.canonicalize().unwrap_or(candidate);
-    candidate.starts_with(&root)
-}
-
-/// True when `path` is a *linked* git worktree rather than the repository's main
-/// checkout. A linked worktree's per-worktree git dir
-/// (`<repo>/.git/worktrees/<name>`) differs from the shared common dir
-/// (`<repo>/.git`), whereas they are identical for the main checkout.
-fn is_linked_worktree(path: &str) -> bool {
-    let git_dir = git_command_stdout(path, &["rev-parse", "--path-format=absolute", "--git-dir"])
-        .ok()
-        .filter(|dir| !dir.is_empty());
-    let common_dir = git_common_dir(path);
-    match (git_dir, common_dir) {
-        (Some(git_dir), Some(common_dir)) => {
-            let git_dir = std::path::PathBuf::from(git_dir);
-            let git_dir = git_dir.canonicalize().unwrap_or(git_dir);
-            !paths_equal(&git_dir, &common_dir)
-        }
-        _ => false,
-    }
-}
-
-/// Resolve the absolute git common dir (the shared repository `.git` directory)
-/// for `path`, canonicalizing so the same repo compares equal regardless of how
-/// it was reached (main checkout vs. linked worktree, symlinks, relative input).
-fn git_common_dir(path: &str) -> Option<std::path::PathBuf> {
-    let common_dir = git_command_stdout(
-        path,
-        &["rev-parse", "--path-format=absolute", "--git-common-dir"],
-    )
-    .ok()
-    .filter(|dir| !dir.is_empty())?;
-    let common_dir = std::path::PathBuf::from(common_dir);
-    Some(common_dir.canonicalize().unwrap_or(common_dir))
-}
-
-/// Compare two filesystem paths, tolerating symlinked/relative differences by
-/// canonicalizing when possible and falling back to a lexical comparison.
-fn paths_equal(a: &std::path::Path, b: &std::path::Path) -> bool {
-    match (a.canonicalize(), b.canonicalize()) {
-        (Ok(ca), Ok(cb)) => ca == cb,
-        _ => a == b,
-    }
-}
-
-/// Reconstruct the [`WorktreeInfo`] for a restored thread worktree path. The
-/// branch is read back from the worktree's HEAD so unmerged-commit / local-change
-/// detection keeps working across restarts.
-///
-/// Returns `None` when the worktree's branch cannot be recovered (detached HEAD
-/// or a failed lookup). Attaching `WorktreeInfo` with an empty branch is unsafe:
-/// idle cleanup builds an `origin/main..<branch>` range from it and, with an
-/// empty branch, inspects the wrong checkout's HEAD — so a clean detached
-/// worktree carrying unmerged work could be wrongly removed (#3011 codex P1).
-fn restored_worktree_info(parent_path: &str, worktree_path: &str) -> Option<WorktreeInfo> {
-    let branch_name = git_command_stdout(worktree_path, &["rev-parse", "--abbrev-ref", "HEAD"])
-        .ok()
-        .filter(|b| !b.is_empty() && b != "HEAD")?;
-    Some(WorktreeInfo {
-        // `original_path` is the cleanup root: idle cleanup runs
-        // `git -C original_path worktree remove <worktree_path>` then
-        // `git -C original_path branch -D <branch>`. It must therefore be a
-        // stable checkout that survives removing this worktree — never the
-        // restored worktree itself (which `parent_path` can equal when the
-        // dispatch parent is itself a linked worktree). Resolve the main
-        // checkout from the worktree's common dir so branch deletion still
-        // works post-removal.
-        original_path: main_checkout_for_worktree(worktree_path)
-            .unwrap_or_else(|| parent_path.to_string()),
-        worktree_path: worktree_path.to_string(),
-        branch_name,
-    })
-}
-
-/// #3207 (part 2): resolve the channel's EXISTING managed worktree for reuse on
-/// the cold-start / resume path, instead of rotating a brand-new
-/// `%Y%m%d-%H%M%S` worktree every turn. claude sessions are scoped to the cwd's
-/// project dir (`~/.claude/projects/<cwd-mangled>/<sid>.jsonl`), so a rotated
-/// worktree makes `--resume` structurally impossible and forces a fresh session
-/// even when the DB still holds a provider session id — the conversation is lost
-/// while the status panel reports "기존 세션 복원". Reusing the prior worktree
-/// keeps the sid's transcript discoverable so the launch genuinely resumes.
-///
-/// This is the SAME persisted mapping the #3011 thread-bootstrap reuse relies on
-/// (`sessions.cwd` keyed by the channel's session-key candidates), with the same
-/// safety filters: the path must be an AgentDesk-managed linked worktree on disk
-/// that belongs to the requested parent repo, with a recoverable branch.
-/// Returns `None` when there is no reusable worktree (genuine fresh start).
-pub(super) fn resolve_reusable_worktree(
-    pg_pool: Option<&sqlx::PgPool>,
-    token_hash: &str,
-    provider: &ProviderKind,
-    channel_name: &str,
-    channel_id: u64,
-    parent_path: &str,
-) -> Option<WorktreeInfo> {
-    let restored = restore_thread_worktree_path_from_db(
-        pg_pool,
-        token_hash,
-        provider,
-        channel_name,
-        channel_id,
-    )
-    .filter(|path| is_managed_worktree_path(path))
-    .filter(|path| restored_worktree_belongs_to_parent(parent_path, path))?;
-    restored_worktree_info(parent_path, &restored)
-}
-
-/// Reconstruct and attach [`WorktreeInfo`] to a restored session when its
-/// `path` is an AgentDesk-managed linked git worktree and the session does not
-/// already carry worktree metadata. No-op otherwise (already populated, a
-/// user-configured workspace outside the managed worktrees root, or the main
-/// checkout). Used by the auto-restore paths so a thread session that resumes
-/// after a dcserver restart regains its worktree metadata, inflight worktree
-/// context, and a stable cleanup root instead of silently dropping them (#3011).
-fn reconstruct_managed_worktree_metadata(
-    session: &mut DiscordSession,
-    provider: &ProviderKind,
-    channel_id: ChannelId,
-    path: &str,
-) {
-    if session.worktree.is_some() || !is_managed_worktree_path(path) || !is_linked_worktree(path) {
-        return;
-    }
-    // Skip reconstruction when no branch can be recovered (detached HEAD); see
-    // `restored_worktree_info` — attaching an empty branch would mislead cleanup.
-    let Some(wt_info) = restored_worktree_info(path, path) else {
-        return;
-    };
-    let base_commit = crate::services::platform::git_head_commit(&wt_info.original_path);
-    sync_inflight_worktree_context(
-        provider,
-        channel_id.get(),
-        Some(wt_info.worktree_path.clone()),
-        Some(wt_info.branch_name.clone()),
-        base_commit,
-    );
-    let ts = chrono::Local::now().format("%H:%M:%S");
-    tracing::info!(
-        "  [{ts}] ↻ Restored worktree metadata: {} (branch: {})",
-        wt_info.worktree_path,
-        wt_info.branch_name
-    );
-    session.worktree = Some(wt_info);
-}
-
-/// Resolve the repository's main checkout directory for a linked `worktree_path`.
-/// The common dir resolves to `<main_checkout>/.git`, so its parent is the main
-/// checkout. Returns `None` when the path is not under a resolvable git repo.
-fn main_checkout_for_worktree(worktree_path: &str) -> Option<String> {
-    let common_dir = git_common_dir(worktree_path)?;
-    let main_checkout = common_dir.parent()?;
-    Some(main_checkout.to_string_lossy().to_string())
-}
-
 /// Create a lightweight session for a thread, bootstrapped from the parent channel's path.
 /// The session's `channel_name` uses `{parent_channel}-t{thread_id}` so the derived
 /// tmux session name stays short and unique instead of using the full thread title.
@@ -1371,21 +423,18 @@ pub(super) async fn bootstrap_thread_session(
         .channel_name
         .clone()
         .unwrap_or_else(|| "unknown".to_string());
-    let restored_worktree = restore_thread_worktree_path_from_db(
+    let restored_worktree = resolve_reusable_worktree(
         shared.pg_pool.as_ref(),
         &shared.token_hash,
         &provider_kind,
         &ch,
         thread_channel_id.get(),
-    )
-    .filter(|path| is_managed_worktree_path(path))
-    .filter(|path| restored_worktree_belongs_to_parent(parent_path, path));
+        parent_path,
+    );
     // Only honor the restore when a branch is recoverable. A detached / unknown
-    // branch would yield an empty `branch_name` that misleads idle cleanup, so
-    // in that case fall through to create a fresh, well-formed worktree instead.
-    if let Some(restored_path) = restored_worktree
-        && let Some(wt_info) = restored_worktree_info(parent_path, &restored_path)
-    {
+    // branch would yield no `WorktreeInfo`, so in that case fall through to
+    // create a fresh, well-formed worktree instead.
+    if let Some(wt_info) = restored_worktree {
         let base_commit = crate::services::platform::git_head_commit(&wt_info.original_path);
         let restored_path = wt_info.worktree_path.clone();
         let ts = chrono::Local::now().format("%H:%M:%S");
@@ -1447,211 +496,6 @@ pub(super) async fn bootstrap_thread_session(
     let ts = chrono::Local::now().format("%H:%M:%S");
     tracing::info!("  [{ts}] ↻ Bootstrapped thread session: {effective_path}");
     true
-}
-
-fn sync_inflight_worktree_context(
-    provider: &crate::services::provider::ProviderKind,
-    channel_id: u64,
-    worktree_path: Option<String>,
-    worktree_branch: Option<String>,
-    base_commit: Option<String>,
-) {
-    if let Some(mut inflight) = super::inflight::load_inflight_state(provider, channel_id) {
-        inflight.set_worktree_context(worktree_path, worktree_branch, base_commit);
-        let _ = super::inflight::save_inflight_state(&inflight);
-    }
-}
-
-/// Resolve the channel name and parent category name for a Discord channel.
-///
-/// `cache` is an optional optimization: when present (leader-side), category
-/// names are looked up via the in-memory guild cache and avoid an extra REST
-/// hop. Worker-side callers without a live shard pass `None` and pay the
-/// REST fallback at line ~978 instead. Correctness is identical either way.
-pub(super) async fn resolve_channel_category(
-    http: &Arc<serenity::http::Http>,
-    cache: Option<&Arc<serenity::cache::Cache>>,
-    channel_id: serenity::model::id::ChannelId,
-) -> (Option<String>, Option<String>) {
-    let Ok(channel) = channel_id.to_channel(http).await else {
-        return (None, None);
-    };
-    let serenity::model::channel::Channel::Guild(gc) = channel else {
-        return (None, None);
-    };
-    let ch_name = Some(gc.name.clone());
-    let cat_name = if let Some(parent_id) = gc.parent_id {
-        let cached_cat_name = cache.and_then(|c| {
-            c.guild(gc.guild_id).and_then(|guild| {
-                guild
-                    .channels
-                    .get(&parent_id)
-                    .map(|parent_ch| parent_ch.name.clone())
-            })
-        });
-
-        if let Some(cat_name) = cached_cat_name {
-            Some(cat_name)
-        } else if let Ok(parent_ch) = parent_id.to_channel(http).await {
-            match parent_ch {
-                serenity::model::channel::Channel::Guild(cat) => Some(cat.name.clone()),
-                _ => {
-                    let ts = chrono::Local::now().format("%H:%M:%S");
-                    tracing::info!(
-                        "  [{ts}] ⚠ Category channel {parent_id} is not a Guild channel for #{}",
-                        gc.name
-                    );
-                    None
-                }
-            }
-        } else {
-            let ts = chrono::Local::now().format("%H:%M:%S");
-            tracing::info!(
-                "  [{ts}] ⚠ Failed to resolve category {parent_id} for #{}",
-                gc.name
-            );
-            None
-        }
-    } else {
-        let ts = chrono::Local::now().format("%H:%M:%S");
-        tracing::info!("  [{ts}] ⚠ No parent_id for #{}", gc.name);
-        None
-    };
-    (ch_name, cat_name)
-}
-
-pub(in crate::services::discord) async fn validate_live_channel_routing_with_dm_hint(
-    ctx: &serenity::prelude::Context,
-    provider: &ProviderKind,
-    settings: &DiscordBotSettings,
-    channel_id: serenity::model::id::ChannelId,
-    is_dm_hint: Option<bool>,
-) -> Result<(), settings::BotChannelRoutingGuardFailure> {
-    let is_dm = match is_dm_hint {
-        Some(is_dm) => is_dm,
-        None => matches!(
-            channel_id.to_channel(&ctx.http).await,
-            Ok(serenity::model::channel::Channel::Private(_))
-        ),
-    };
-    let (channel_name, _) = resolve_channel_category(&ctx.http, Some(&ctx.cache), channel_id).await;
-    let (allowlist_channel_id, provider_channel_name) = if let Some((parent_id, parent_name)) =
-        resolve_thread_parent(&ctx.http, channel_id).await
-    {
-        (parent_id, parent_name.or(channel_name.clone()))
-    } else {
-        (channel_id, channel_name.clone())
-    };
-    validate_bot_channel_routing_with_provider_channel(
-        settings,
-        provider,
-        allowlist_channel_id,
-        channel_name.as_deref(),
-        provider_channel_name.as_deref(),
-        is_dm,
-    )
-}
-
-pub(in crate::services::discord) async fn validate_live_channel_routing(
-    ctx: &serenity::prelude::Context,
-    provider: &ProviderKind,
-    settings: &DiscordBotSettings,
-    channel_id: serenity::model::id::ChannelId,
-) -> Result<(), settings::BotChannelRoutingGuardFailure> {
-    validate_live_channel_routing_with_dm_hint(ctx, provider, settings, channel_id, None).await
-}
-
-pub(super) async fn provider_handles_channel(
-    ctx: &serenity::prelude::Context,
-    provider: &ProviderKind,
-    settings: &DiscordBotSettings,
-    channel_id: serenity::model::id::ChannelId,
-) -> bool {
-    validate_live_channel_routing(ctx, provider, settings, channel_id)
-        .await
-        .is_ok()
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(super) enum RuntimeChannelBindingStatus {
-    Owned,
-    Unowned,
-    Unknown,
-}
-
-pub(super) async fn resolve_runtime_channel_binding_status(
-    http: &Arc<serenity::Http>,
-    channel_id: serenity::model::id::ChannelId,
-) -> RuntimeChannelBindingStatus {
-    if settings::has_configured_channel_binding(channel_id, None) {
-        return RuntimeChannelBindingStatus::Owned;
-    }
-
-    let Ok(channel) = channel_id.to_channel(http).await else {
-        return RuntimeChannelBindingStatus::Unknown;
-    };
-
-    match channel {
-        serenity::model::channel::Channel::Private(_) => RuntimeChannelBindingStatus::Owned,
-        serenity::model::channel::Channel::Guild(gc) => {
-            use poise::serenity_prelude::ChannelType;
-            match gc.kind {
-                ChannelType::PublicThread | ChannelType::PrivateThread => {
-                    let Some(parent_id) = gc.parent_id else {
-                        return RuntimeChannelBindingStatus::Unowned;
-                    };
-                    let parent_name = match parent_id.to_channel(http).await {
-                        Ok(serenity::model::channel::Channel::Guild(parent)) => {
-                            Some(parent.name.clone())
-                        }
-                        Ok(_) => None,
-                        Err(_) => None,
-                    };
-                    if settings::has_configured_channel_binding(parent_id, parent_name.as_deref()) {
-                        RuntimeChannelBindingStatus::Owned
-                    } else {
-                        RuntimeChannelBindingStatus::Unowned
-                    }
-                }
-                _ => {
-                    if settings::has_configured_channel_binding(channel_id, Some(&gc.name)) {
-                        RuntimeChannelBindingStatus::Owned
-                    } else {
-                        RuntimeChannelBindingStatus::Unowned
-                    }
-                }
-            }
-        }
-        _ => RuntimeChannelBindingStatus::Unowned,
-    }
-}
-
-/// If `channel_id` is a Discord thread, return the parent channel ID and name.
-/// For non-thread channels, returns `None`.
-pub(super) async fn resolve_thread_parent(
-    http: &Arc<serenity::Http>,
-    channel_id: serenity::model::id::ChannelId,
-) -> Option<(serenity::model::id::ChannelId, Option<String>)> {
-    let channel = channel_id.to_channel(http).await.ok()?;
-    let serenity::model::channel::Channel::Guild(gc) = channel else {
-        return None;
-    };
-    use poise::serenity_prelude::ChannelType;
-    match gc.kind {
-        ChannelType::PublicThread | ChannelType::PrivateThread => {
-            let parent_id = gc.parent_id?;
-            let parent_name = if let Ok(parent_ch) = parent_id.to_channel(http).await {
-                match parent_ch {
-                    serenity::model::channel::Channel::Guild(pg) => Some(pg.name.clone()),
-                    _ => None,
-                }
-            } else {
-                None
-            };
-            Some((parent_id, parent_name))
-        }
-        _ => None,
-    }
 }
 
 #[cfg(test)]
